requirements.txt
pandas
openpyxl
pyyaml
SQLAlchemy
psycopg2-binary
